% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{absolute_from}
\alias{absolute_from}
\title{Computation of absolute from spillovers at time T}
\usage{
absolute_from(listAtT)
}
\arguments{
\item{listAtT}{takes in a list containing (G)FEVD decomposed at frequencies at time t}
}
\value{
list with absolute from spillovers at frequencies
}
\description{
Computation of absolute from spillovers at time T
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
