% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{absolute_pairwise}
\alias{absolute_pairwise}
\title{Computation of absolute pairwise spillovers at time T}
\usage{
absolute_pairwise(listAtT)
}
\arguments{
\item{listAtT}{takes in a list containing (G)FEVD decomposed at frequencies at time t}
}
\value{
list with absolute pairwise spillovers at frequencies
}
\description{
Computation of absolute pairwise spillovers at time T
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
