% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{absolute_to}
\alias{absolute_to}
\title{Computation of absolute to spillovers at time T}
\usage{
absolute_to(listAtT)
}
\arguments{
\item{listAtT}{takes in a list containing (G)FEVD decomposed at frequencies at time t}
}
\value{
list with absolute to spillovers at frequencies
}
\description{
Computation of absolute to spillovers at time T
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
