% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{collapseAll}
\alias{collapseAll}
\title{Collapsing all bounds}
\usage{
collapseAll(estimate)
}
\arguments{
\item{estimate}{the object of class connectedness_estimate}
}
\value{
object of class connectedness_estimate
}
\description{
Collapsing all bounds
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
