% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of_spills_methods.R
\name{collapseBounds.list_of_spills}
\alias{collapseBounds.list_of_spills}
\title{Function to collapse bounds}
\usage{
\method{collapseBounds}{list_of_spills}(spillover_table, which)
}
\arguments{
\item{spillover_table}{a list_of_spills object, ideally from the provided estimation 
functions}

\item{which}{which frequency bands to collapse. Should be a sequence like 1:2
or 1:5, etc.}

\item{...}{for the sake of CRAN not to complain}
}
\value{
list_of_spills with less frequency bands.
}
\description{
Taking in list_of_spills, if the individual spillover_tables are frequency 
based, it allows you to collapse several frequency bands into one.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
