% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover_table_methods.R
\name{collapseBounds.spillover_table}
\alias{collapseBounds.spillover_table}
\title{Function to collapse bounds}
\usage{
\method{collapseBounds}{spillover_table}(x, which)
}
\arguments{
\item{x}{a spillover_table object, ideally from the provided estimation 
functions}

\item{which}{which frequency bands to collapse. Should be a sequence like 1:2
or 1:5, etc.}
}
\value{
spillover_table with less frequency bands.
}
\description{
Taking in spillover_table, if the spillover_table is frequency based, it 
allows you to collapse several frequency bands into one.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
