% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spilloverTable.R
\name{createSpilloverTable}
\alias{createSpilloverTable}
\title{Create a latex spillover table like in Barunik, Krehlik (2015)}
\usage{
createSpilloverTable(tot, decomposition_a, decomposition_r, nams,
  format = "\%1.1f", sep = "-")
}
\arguments{
\item{tot}{a matrix usually from spilloverDY12 or spilloverDY09}

\item{decomposition_a}{a list of matrices containing absolute decomposition typically from spilloverBK12}

\item{decomposition_r}{a list of matrices containing within decomposition typically from spilloverBK12}

\item{nams}{names of the variables}

\item{format}{a format of the output numbers}

\item{sep}{a separator in the decomposition}
}
\description{
Create a latex spillover table like in Barunik, Krehlik (2015)
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
