% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spilloversRolling.R
\docType{data}
\name{exampleSim}
\alias{exampleSim}
\title{The simulated time-series}
\description{
The dataset includes three simulated processes with spillover dynamics.
}
\author{
Tomas Krehlik \email{tomas.krehlik@gmail.com}
}
\keyword{data}
