% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevd.R
\name{fevd}
\alias{fevd}
\title{Compute a forecast error vector decomposition in recursive identification scheme}
\usage{
fevd(est, n.ahead = 100, no.corr = F)
}
\arguments{
\item{est}{the VAR estimate from the vars package}

\item{n.ahead}{how many periods ahead should be taken into account}

\item{no.corr}{boolean if the off-diagonal elements should be set to 0.}
}
\value{
a matrix that corresponds to contribution of ith variable to jth variance of forecast
}
\description{
This function computes the standard forecast error vector decomposition given the 
estimate of the VAR.
}
\author{
Tomas Krehlik \email{tomas.krehlik@gmail.com}
}
