% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevd.R
\name{fftGenFEVD}
\alias{fftGenFEVD}
\title{Compute a FFT transform of forecast error vector decomposition in generalised VAR scheme.}
\usage{
fftGenFEVD(est, n.ahead = 100, no.corr = F, range)
}
\arguments{
\item{est}{the VAR estimate from the vars package}

\item{n.ahead}{how many periods ahead should be taken into account}

\item{no.corr}{boolean if the off-diagonal elements should be set to 0.}

\item{range}{defines the frequency partitions to which the spillover should be decomposed}
}
\value{
a list of matrices that corresponds to contribution of ith variable to jth variance of forecast
}
\description{
This function computes the decomposition of standard forecast error vector decomposition given the 
estimate of the VAR. The decomposition is done according to the Stiassny (1996)
}
\author{
Tomas Krehlik \email{tomas.krehlik@gmail.com}
}
