% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover_table_methods.R
\name{from.spillover_table}
\alias{from.spillover_table}
\title{Function to compute from spillovers}
\usage{
\method{from}{spillover_table}(spillover_table, within = F, ...)
}
\arguments{
\item{within}{whether to compute the within spillovers if the spillover
tables are frequency based.}

\item{...}{for the sake of CRAN not to complain}

\item{x}{a spillover_table object, ideally from the provided estimation 
functions}
}
\value{
a list containing the from spillover
}
\description{
Taking in spillover_table, the function computes the from spillover.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
