% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{fromSpillovers}
\alias{fromSpillovers}
\title{Compute from others spillovers from the estimates}
\usage{
fromSpillovers(estimate)
}
\arguments{
\item{estimate}{an outcome from either frequency based or standard connectedness measures}
}
\value{
either a vector or list of from others spillovers depending on input.
}
\description{
This function does the computation of the from others spillovers from the
estimate of connectedness table
}
\author{
Tomas Krehlik \email{tomas.krehlik@sorgmail.com}
}
