% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{get_absolute_to}
\alias{get_absolute_to}
\title{Computation of absolute to spillovers at time T}
\usage{
get_absolute_to(ce)
}
\arguments{
\item{ce}{takes in a structure of class connectedness_estimate}
}
\value{
list with absolute to spillovers at frequencies
}
\description{
Computation of absolute to spillovers at time T
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
