% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{get_type}
\alias{get_type}
\title{Computation of a given type of spillover}
\usage{
get_type(f, ce)
}
\arguments{
\item{f}{function that is specific for the type}

\item{ce}{the object of class connectedness_estimate}
}
\value{
list with within from spillovers at frequencies, useful for plotting
}
\description{
Computation of a given type of spillover
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
