% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{get_within_net}
\alias{get_within_net}
\title{Computation of within net spillovers at time T}
\usage{
get_within_net(ce)
}
\arguments{
\item{ce}{takes in a structure of class connectedness_estimate}
}
\value{
list with within net spillovers at frequencies
}
\description{
Computation of within net spillovers at time T
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
