% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{get_within_pairwise}
\alias{get_within_pairwise}
\title{Computation of within pairwise spillovers at time T}
\usage{
get_within_pairwise(ce)
}
\arguments{
\item{ce}{takes in a structure of class connectedness_estimate}
}
\value{
list with within pairwise spillovers at frequencies
}
\description{
Computation of within pairwise spillovers at time T
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
