% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getIndeces}
\alias{getIndeces}
\title{Get the indeces for the individual intervals}
\usage{
getIndeces(n.ahead, up, down)
}
\arguments{
\item{n.ahead}{the length of the vector coming out of the DFT}

\item{up}{the upper boundary of the interval}

\item{down}{the lower boundary of the interval}
}
\description{
This function returns the indeces of the vector coming from DFT of time series
of length n.ahead that correspond to frequencies in the interval (up, down].
}
\author{
Tomas Krehlik \email{tomas.krehlik@sorgmail.com}
}
