% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getPartition}
\alias{getPartition}
\title{Get a list of indeces corresponding to parts of frequency partition}
\usage{
getPartition(partition, n.ahead)
}
\arguments{
\item{partition}{breaking points of partition of frequency interval, should be ordered decreasingly.}

\item{n.ahead}{how many observations is the FFT done on.}
}
\value{
a list of vectors of indeces corresponding to individual partitions
}
\description{
This function takes in a vector of numbers denoting the breaks in partition of an interval
and returns a list of indeces that correspond to indeces that are contained within an individual
intervals. The individual parts then contain (a,b] for all pairs in the interval. Hence if you
want pi to be included, the partition should start with something slightly bigger than pi.
}
\author{
Tomas Krehlik \email{tomas.krehlik@sorgmail.com}
}
