% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{makeStructure}
\alias{makeStructure}
\title{Making structured object}
\usage{
makeStructure(estimate, dates, bounds)
}
\arguments{
\item{estimate}{a list of lists of (G)FEVD tables on given frequency in a given day.}

\item{dates}{the dates vector of the same length as the estimate object}

\item{bounds}{numeric vector giving bounds of the frequency bounds.}
}
\value{
object of class connectedness_estimate
}
\description{
The function returns a structured object that is useful for plotting
and extracting other types of spillovers, such as TO, FROM, etc. types
of spillovers. The same structure is also the output of rolled BK spillovers
estimates.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
