% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover_table_methods.R
\name{net.spillover_table}
\alias{net.spillover_table}
\title{Function to compute net spillovers}
\usage{
\method{net}{spillover_table}(spillover_table, within = F, ...)
}
\arguments{
\item{spillover_table}{a spillover_table object, ideally from the provided estimation 
functions}

\item{within}{whether to compute the within spillovers if the spillover
tables are frequency based.}

\item{...}{for the sake of CRAN not to complain}
}
\value{
a list containing the net spillover
}
\description{
Taking in spillover_table, the function computes the net spillover.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
