% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of_spills_methods.R
\name{overall.list_of_spills}
\alias{overall.list_of_spills}
\title{Function to compute overall spillovers}
\usage{
\method{overall}{list_of_spills}(spillover_table, within = F, ...)
}
\arguments{
\item{spillover_table}{a list_of_spills object, ideally from rolling window estimation}

\item{within}{whether to compute the within spillovers if the spillover
tables are frequency based.}

\item{...}{for the sake of CRAN not to complain}
}
\value{
a list containing the overall spillovers
}
\description{
Taking in list_of_spillovers, the function computes the overall spillovers
for all the individual spillover_table.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
