% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers_exported_methods.R
\name{overall}
\alias{overall}
<<<<<<< HEAD
\title{Function to compute overall spillovers}
\usage{
overall(x, ...)
}
\arguments{
\item{x}{either the spillover_table or list_of_spills object.}

\item{...}{this argument can contain within = F or T as to whether to 
compute the within spillovers if the spillover tables are frequency based.}
}
\value{
a list containing the overall spillover
}
\description{
Taking in spillover_table or list_of spills object, the function computes 
the overall spillover.
=======
\title{Method for computing overall spillovers}
\usage{
overall(spillover_table, ...)
}
\arguments{
\item{spillover_table}{the output of spillover estimation function 
or rolling spillover estimation function}

\item{...}{other arguments like whether it is within or absolute spillover
in case of the frequency spillovers}
}
\value{
Value for overall spillover
}
\description{
Method for computing overall spillovers
>>>>>>> BigVAR
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
