% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of_spills_methods.R
\name{pairwise.list_of_spills}
\alias{pairwise.list_of_spills}
\title{Function to compute pairwise spillovers}
\usage{
\method{pairwise}{list_of_spills}(x, within = F)
}
\arguments{
\item{x}{a list_of_spills object, ideally from rolling window estimation}

\item{within}{whether to compute the within spillovers if the spillover
tables are frequency based.}
}
\value{
a list containing the pairwise spillovers
}
\description{
Taking in list_of_spillovers, the function computes the pairwise spillovers
for all the individual spillover_table.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
