% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers_exported_methods.R
\name{pairwise}
\alias{pairwise}
\title{Function to compute overall spillovers}
\usage{
pairwise(x, ...)
}
\arguments{
\item{x}{either the spillover_table or list_of_spills object.}

\item{...}{this argument can contain within = F or T as to whether to 
compute the within spillovers if the spillover tables are frequency based.}
}
\value{
a list containing the overall spillover
}
\description{
Taking in spillover_table or list_of spills object, the function computes 
the overall spillover.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
