% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{plot_spills}
\alias{plot_spills}
\title{Function that plots various types of spillovers}
\usage{
plot_spills(s, otherAttributes = NULL)
}
\arguments{
\item{s}{a list of named matrices with spillovers}

\item{otherAttributes}{other attributes that can be used to enhance plot}
}
\description{
Function that plots various types of spillovers
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
