% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of_spills_methods.R
\name{plotNet.list_of_spills}
\alias{plotNet.list_of_spills}
\title{Function to plot net spillovers}
\usage{
\method{plotNet}{list_of_spills}(x, within = F,
  which = 1:nrow(x$list_of_tables[[1]]$tables[[1]]))
}
\arguments{
\item{x}{a list_of_spills object, ideally from rolling window estimation}

\item{within}{whether to compute the within spillovers if the spillover
tables are frequency based.}
}
\value{
a plot of net spillovers
}
\description{
Taking in list_of_spillovers, the function plots the net spillovers
using the plot.zoo function
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
