% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of_spills_methods.R
\name{plotOverall.list_of_spills}
\alias{plotOverall.list_of_spills}
\title{Function to plot overall spillovers}
\usage{
\method{plotOverall}{list_of_spills}(x, within = F)
}
\arguments{
\item{x}{a list_of_spills object, ideally from rolling window estimation}

\item{within}{whether to compute the within spillovers if the spillover
tables are frequency based.}
}
\value{
a plot of overall spillovers
}
\description{
Taking in list_of_spillovers, the function plots the overall spillovers
using the plot.zoo function
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
