% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of_spills_methods.R
\name{plotSpecific.list_of_spills}
\alias{plotSpecific.list_of_spills}
\title{Function to plot specific spilover from i to j}
\usage{
\method{plotSpecific}{list_of_spills}(spillover_table, i, j, ...)
}
\arguments{
\item{spillover_table}{a list_of_spills object, ideally from rolling window estimation}

\item{i}{from variable}

\item{j}{to variable}

\item{...}{for the sake of CRAN not to complain}
}
\value{
a plot of pairwise spillovers
}
\description{
Taking in list_of_spillovers, the function plots the spillover from i to j
using the zoo::plot.zoo function
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
