% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers_exported_methods.R
\name{plotSpecific}
\alias{plotSpecific}
\title{Method for ploting specific pair spillover}
\usage{
plotSpecific(spillover_table, ...)
}
\arguments{
\item{spillover_table}{the output of rolling spillover estimation function}

\item{...}{other arguments like which specifi pair to plot.}
}
\value{
The plot
}
\description{
Method for ploting specific pair spillover
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
