% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{plotSpills}
\alias{plotSpills}
\title{Function for convenient plotting}
\usage{
plotSpills(ce, which = "absolute", type = "to", otherAttributes = NULL)
}
\arguments{
\item{ce}{takes in a structure of class connectedness_estimate}

\item{which}{is either "absolute" or "within"}

\item{type}{is one of the following: to, from, net, pairwise.}

\item{otherAttributes}{other attributes that can be used to enhance plot}
}
\description{
Function for convenient plotting
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
