% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{plotSpillsAll}
\alias{plotSpillsAll}
\title{Function for convenient plotting}
\usage{
plotSpillsAll(ce, file, otherAttributes = NULL)
}
\arguments{
\item{ce}{takes in a structure of class connectedness_estimate}

\item{otherAttributes}{other attributes that can be used to enhance plot}

\item{which}{is either "absolute" or "within"}

\item{type}{is one of the following: to, from, net, pairwise.}
}
\description{
This function takes in a path to a file and plots all possible combinations
of spillovers.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}

