% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers_exported_methods.R
\name{plotTo}
\alias{plotTo}
<<<<<<< HEAD
\title{Function to plot to spillovers}
\usage{
plotTo(x, ...)
}
\arguments{
\item{x}{a list_of_spills object, ideally from rolling window estimation}

\item{...}{contains two arguments within and which, for details see mtehod
plotTo.list_of_spills}
}
\value{
a plot of to spillovers
}
\description{
Taking in list_of_spillovers, the function plots the to spillovers
using the zoo::plot.zoo function
=======
\title{Method for ploting TO spillovers}
\usage{
plotTo(spillover_table, ...)
}
\arguments{
\item{spillover_table}{the output of rolling spillover estimation function}

\item{...}{other arguments like whether it is within or absolute spillover
in case of the frequency spillovers}
}
\value{
The plot
}
\description{
Method for ploting TO spillovers
>>>>>>> BigVAR
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
