% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of_spills_methods.R
\name{print.list_of_spills}
\alias{print.list_of_spills}
\title{Function to not print the list_of_spills object}
\usage{
\method{print}{list_of_spills}(x, ...)
}
\arguments{
\item{x}{a list_of_spills object, ideally from the provided estimation 
functions}

\item{...}{for the sake of CRAN not to complain}
}
\description{
Usually it is not a good idea to print the list_of_spills object, hence
this function implements warning and shows how to print them individually
if the user really wants to.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
