% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover_table_methods.R
\name{print.spillover_table}
\alias{print.spillover_table}
\title{Function to print the spillover table object}
\usage{
\method{print}{spillover_table}(x, ...)
}
\arguments{
\item{x}{a spillover_table object, ideally from the provided estimation 
functions}

\item{...}{for the sake of CRAN not to complain}
}
\description{
The function takes as an argument the spillover_table object and prints it
nicely to the console. While doing that it also computes all the neccessary
measures.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
