% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spillovers.R
\name{spilloverRollingDY12}
\alias{spilloverRollingDY12}
\title{Computing rolling spillover from the generalized fevd according to Diebold Yilmaz (2012)}
\usage{
spilloverRollingDY12(data, p, type, window, n.ahead, table = F, no.corr,
  cluster = NULL)
}
\arguments{
\item{data}{variable containing the dataset}

\item{p}{lags in the VAR estimate.}

\item{type}{which type of VAR to use, see help for VAR from vars package}

\item{window}{length of the window to be rolled}

\item{n.ahead}{how many periods ahead should the FEVD be computed, generally this number
should be high enough so that it won't change with additional period}

\item{table}{boolean whether the full spillover table should be returned}

\item{no.corr}{boolean parameter whether the off-diagonal in the covariance matrix should be
set to zero}

\item{cluster}{either NULL for no parallel processing or the variable containing the cluster.}
}
\description{
This function computes the rolling spillover using the standard VAR estimate.
We implement the parallel version for faster processing. The window is of fixed window
and is rolled over the data. Interpretation of the other parameters is the same as in the
standard computation of spillover.
}
\examples{
w<-c(0,0)
C<-rbind(c(1,0.0),c(0.0,1))
A1<-rbind(c(0.9,0.09),c(0.09,0.9))
A2<-rbind(c(0.9,0.03),c(0.03,0.9))
data <- rbind(mAr.sim(w,A1,C,N=1000), mAr.sim(w,A2,C,N=1000))
plot(spilloverRollingDY12(data, p = 1, type = "const", window = 200, n.ahead = 100, table = F, no.corr = F), type="l")
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}

