% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spilloversRolling.R
\docType{data}
\name{volatilities}
\alias{volatilities}
\title{Volatilities from Ox Man Institute}
\description{
The dataset includes median realised volatilities of some financial indices
}
\author{
Tomas Krehlik \email{tomas.krehlik@gmail.com}
}
\keyword{data}
