% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{within_from}
\alias{within_from}
\title{Computation of within from spillovers at time T}
\usage{
within_from(listAtT)
}
\arguments{
\item{listAtT}{takes in a list containing (G)FEVD decomposed at frequencies at time t}
}
\value{
list with within from spillovers at frequencies
}
\description{
Computation of within from spillovers at time T
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
