% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{within_pairwise}
\alias{within_pairwise}
\title{Computation of within pairwise spillovers at time T}
\usage{
within_pairwise(listAtT)
}
\arguments{
\item{listAtT}{takes in a list containing (G)FEVD decomposed at frequencies at time t}
}
\value{
list with within pairwise spillovers at frequencies
}
\description{
Computation of within pairwise spillovers at time T
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
