% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{within_to}
\alias{within_to}
\title{Computation of within to spillovers at time T}
\usage{
within_to(listAtT)
}
\arguments{
\item{listAtT}{takes in a list containing (G)FEVD decomposed at frequencies at time t}
}
\value{
list with within to spillovers at frequencies
}
\description{
Computation of within to spillovers at time T
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
