% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{collapseBounds}
\alias{collapseBounds}
\title{Collapsing bounds into less elements}
\usage{
collapseBounds(estimate, index_start, index_end)
}
\arguments{
\item{estimate}{the object of class connectedness_estimate}

\item{index_start}{from where to start collapsing}

\item{index_end}{where to end the collapsing}
}
\value{
object of class connectedness_estimate
}
\description{
Given that your estimate is saved as an object from makeStructure function
this function allows you to collapse the estimates from several bounds. Esentially,
the GFEVD is linear in frequencies so we can just sum up parts of the object.
It is especially useful, when we have estimated too many bounds and want to use only
several of them. This function is especially useful for plotting. See github for examples.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
