% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers_exported_methods.R
\name{collapseBounds}
\alias{collapseBounds}
\title{Method for for collapsing bound for frequency spillovers}
\usage{
collapseBounds(spillover_table, which)
}
\arguments{
\item{spillover_table}{the output of spillover estimation function 
or rolling spillover estimation function}

\item{which}{integer vector indicating which of the frequency bounds
we want to have collapsed}
}
\value{
New spillover object with collapsed bounds
}
\description{
Method for for collapsing bound for frequency spillovers
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
