% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers_exported_methods.R
\name{from}
\alias{from}
\title{Method for computing FROM spillovers}
\usage{
from(spillover_table, ...)
}
\arguments{
\item{spillover_table}{the output of spillover estimation function 
or rolling spillover estimation function}

\item{...}{other arguments like whether it is within or absolute spillover
in case of the frequency spillovers}
}
\value{
Value for FROM spillover
}
\description{
Method for computing FROM spillovers
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
