% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers_exported_methods.R
\name{pairwise}
\alias{pairwise}
\title{Method for computing PAIRWISE spillovers}
\usage{
pairwise(spillover_table, ...)
}
\arguments{
\item{spillover_table}{the output of spillover estimation function 
or rolling spillover estimation function}

\item{...}{other arguments like whether it is within or absolute spillover
in case of the frequency spillovers}
}
\value{
Value for PAIRWISE spillover
}
\description{
Method for computing PAIRWISE spillovers
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
