% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers_exported_methods.R
\name{plotNet}
\alias{plotNet}
\title{Method for ploting NET spillovers}
\usage{
plotNet(spillover_table, ...)
}
\arguments{
\item{spillover_table}{the output of rolling spillover estimation function}

\item{...}{other arguments like whether it is within or absolute spillover
in case of the frequency spillovers}
}
\value{
The plot
}
\description{
Method for ploting NET spillovers
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
