% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers_exported_methods.R
\name{plotPairwise}
\alias{plotPairwise}
<<<<<<< HEAD
\title{Function to plot pairwise spillovers}
\usage{
plotPairwise(x, ...)
}
\arguments{
\item{x}{a list_of_spills object, ideally from rolling window estimation}

\item{...}{contains two arguments within and which, for details see mtehod
plotPairwise.list_of_spills}
}
\value{
a plot of pairwise spillovers
}
\description{
Taking in list_of_spillovers, the function plots the pairwise spillovers
using the zoo::plot.zoo function
=======
\title{Method for ploting PAIRWISE spillovers}
\usage{
plotPairwise(spillover_table, ...)
}
\arguments{
\item{spillover_table}{the output of rolling spillover estimation function}

\item{...}{other arguments like whether it is within or absolute spillover
in case of the frequency spillovers}
}
\value{
The plot
}
\description{
Method for ploting PAIRWISE spillovers
>>>>>>> BigVAR
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
