% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalSpillovers.R
\name{plotSpillsAll}
\alias{plotSpillsAll}
\title{Function for convenient plotting}
\usage{
plotSpillsAll(ce, file, otherAttributes = NULL, height = 7, width = 7)
}
\arguments{
\item{ce}{takes in a structure of class connectedness_estimate}

\item{file}{where the final plot should go}

\item{otherAttributes}{other attributes that can be used to enhance plot}

\item{height}{height of the plot}

\item{width}{width of the plot}
}
\description{
This function takes in a path to a file and plots all possible combinations
of spillovers.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
