% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers_exported_methods.R
\name{plotTo}
\alias{plotTo}
\title{Function to plot to spillovers}
\usage{
plotTo(x, ...)
}
\arguments{
\item{x}{a list_of_spills object, ideally from rolling window estimation}

\item{...}{contains two arguments within and which, for details see mtehod
plotTo.list_of_spills}
}
\value{
a plot of to spillovers
}
\description{
Taking in list_of_spillovers, the function plots the to spillovers
using the zoo::plot.zoo function
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
