% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers.R
\name{spilloverBK09}
\alias{spilloverBK09}
\title{Computing the decomposed spillover from a fevd as defined by Barunik, Krehlik (2018)}
\usage{
spilloverBK09(est, n.ahead = 100, no.corr, partition)
}
\arguments{
\item{est}{the estimate of a system, typically VAR estimate in our case}

\item{n.ahead}{how many periods ahead should the FEVD be computed, generally this number
should be high enough so that it won't change with additional period}

\item{no.corr}{boolean parameter whether the off-diagonal in the covariance matrix should be
set to zero}

\item{partition}{defines the frequency partitions to which the spillover should be decomposed}
}
\value{
spillover_table object
}
\description{
This function is an internal implementation of the frequency spillover.
We apply the identification scheme suggested by fevd to the frequency
decomposition of the transfer functions from the estimate est.
}
\author{
Tomas Krehlik <tomas.krehlik@gmail.com>
}
